Safetly seed languages. Log to stdout in development.

+++ db/seeds.rb
@@ -282,9 +282,15 @@ languages = [
   }
 ]
 
+if Language.count && ENV["DISABLE_DATABASE_ENVIRONMENT_CHECK"] == "1"
+  ActiveRecord::Base.transaction do
+    Language.delete_all
+    ActiveRecord::Base.connection.reset_pk_sequence!('languages')
+  end
+end
 
-ActiveRecord::Base.transaction do
-  Language.delete_all
-  ActiveRecord::Base.connection.reset_pk_sequence!('languages')
-  Language.create(languages)
+unless Language.count
+  ActiveRecord::Base.transaction do
+    Language.create(languages)
+  end
 end
